Update README.md
This is a personal tool for helping my partner at the time and I get the easter egg on Voyage of Destruction for BLOPs4. It made more sense at the time to write an entire application in C++ instead of making paper notes every time 😅 Written using the Qt framework. It's not 100% complete because we couldn't make it past a certain level and later, I found an already completed web app doing the exact same thing. Hah, I have no regrets.